get building on doze
git-svn-id: svn+ssh://corpsvnproxy/yahoo/BrowserPlus/public_platform/trunk@16 3fcf768e-b076-0410-b9b8-e8c34e2d470b

+++ src/libs/bputils/src/Platform/ServiceInterfaceCache.cpp
@@ -53,7 +53,7 @@ bp::serviceInterfaceCache::isNewerThan(const std::string & name,
 
     bp::file::Path path = buildPath(name, version);    
     bool rval = false;
-    BPTime pathTime(0);
+    BPTime pathTime((long)0);
     if (boost::filesystem::exists(path)) {
         try {
             pathTime.set(boost::filesystem::last_write_time(path));
